+++ changelog.md
@@ -3,6 +3,7 @@
 - add context menu to invoice table
 - clients should be also sortable
 - all clients should be loaded, otherwise client selection list all options.
+- when updating invoice details, the corresponging filtered pages should also be updated. 
 - Refactor/organize based on [slide for Week 5 Office Hour 1](https://docs.google.com/presentation/d/1pLZ8Zvpyr_myotpvyVYzY5awpUFUO78DE5CnBnVNm7c/edit#slide=id.g117828bad90_0_9)
 - Reemplazar hamburget con speed dial https://mui.com/components/speed-dial/
 
@@ -16,6 +17,7 @@
 # 2022 03 19
 - Add support in invoice store slice to filter and sort invoices.
 - Show more columns in the invoices pages
+- Fix total invoices for logged user
 # 2022 03 18
 - created ui for filtering and sorting invoices